toolkit/components/url-classifier/tests/mochitest/test_trackingprotection_whitelist.html

<!DOCTYPE HTML>
<!-- Any copyright is dedicated to the Public Domain.
     http://creativecommons.org/publicdomain/zero/1.0/ -->
<html>
<head>
  <title>Test Tracking Protection in Private Browsing mode</title>
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
</head>

<body>
<p id="display"></p>
<div id="content" style="display: none">
</div>
<pre id="test">

<script class="testbody" type="text/javascript">

var mainWindow = window.docShell.rootTreeItem.domWindow;
var contentPage1 = "http://www.itisatrap.org/tests/toolkit/components/url-classifier/tests/mochitest/whitelistFrame.html";
var contentPage2 = "http://example.com/tests/toolkit/components/url-classifier/tests/mochitest/whitelistFrame.html";

ChromeUtils.import("resource://gre/modules/Services.jsm");
const {UrlClassifierTestUtils} = ChromeUtils.import("resource://testing-common/UrlClassifierTestUtils.jsm");
const {BrowserTestUtils} = ChromeUtils.import("resource://testing-common/BrowserTestUtils.jsm");
const {TestUtils} = ChromeUtils.import("resource://testing-common/TestUtils.jsm");

function testOnWindow(contentPage) {
  return new Promise((resolve, reject) => {
    var win = mainWindow.OpenBrowserWindow();
    win.addEventListener("load", function() {
      TestUtils.topicObserved("browser-delayed-startup-finished",
                              subject => subject == win).then(() => {
        win.addEventListener("DOMContentLoaded", function onInnerLoad() {
          if (win.content.location.href != contentPage) {
            BrowserTestUtils.loadURI(win.gBrowser, contentPage);
            return;
          }
          win.removeEventListener("DOMContentLoaded", onInnerLoad, true);

          win.content.addEventListener("load", function innerLoad2() {
            win.content.removeEventListener("load", innerLoad2);
            SimpleTest.executeSoon(function() {
              resolve(win);
            });
          }, false, true);
        }, true);
        SimpleTest.executeSoon(function() {
          BrowserTestUtils.loadURI(win.gBrowser, contentPage);
        });
      });
    }, {capture: true, once: true});
  });
}

var alwaysbadids = [
  "badscript",
];

function checkLoads(aWindow, aWhitelisted, tpEnabled) {
  var win = aWindow.content;
  if (!tpEnabled) {
    is(win.document.getElementById("badscript").dataset.touched, "yes", "Should load tracking javascript");
    is(win.document.blockedNodeByClassifierCount, 0, "Should not identify any tracking elements");
    return;
  }

  is(win.document.getElementById("badscript").dataset.touched, "no", "Should not load tracking javascript");
  is(win.document.getElementById("goodscript").dataset.touched, aWhitelisted ? "yes" : "no", "Should load whitelisted tracking javascript");

  var badids = alwaysbadids.slice();
  if (!aWhitelisted) {
    badids.push("goodscript");
  }
  is(win.document.blockedNodeByClassifierCount, badids.length, "Should identify all tracking elements");

  var blockedNodes = win.document.blockedNodesByClassifier;

  // Make sure that every node in blockedNodes exists in the tree
  // (that may not always be the case but do not expect any nodes to disappear
  // from the tree here)
  var allNodeMatch = true;
  for (let i = 0; i < blockedNodes.length; i++) {
    let nodeMatch = false;
    for (let j = 0; j < badids.length && !nodeMatch; j++) {
      nodeMatch = nodeMatch ||
        (blockedNodes[i] == win.document.getElementById(badids[j]));
    }

    allNodeMatch = allNodeMatch && nodeMatch;
  }
  is(allNodeMatch, true, "All annotated nodes are expected in the tree");

  // Make sure that every node with a badid (see badids) is found in the
  // blockedNodes. This tells us if we are neglecting to annotate
  // some nodes
  allNodeMatch = true;
  for (let j = 0; j < badids.length; j++) {
    let nodeMatch = false;
    for (let i = 0; i < blockedNodes.length && !nodeMatch; i++) {
      nodeMatch = nodeMatch ||
        (blockedNodes[i] == win.document.getElementById(badids[j]));
    }

    allNodeMatch = allNodeMatch && nodeMatch;
  }
  is(allNodeMatch, true, "All tracking nodes are expected to be annotated as such");
}

SpecialPowers.pushPrefEnv(
  {"set": [["privacy.trackingprotection.enabled", true],
           ["channelclassifier.allowlist_example", true]]},
  test);

async function test() {
  SimpleTest.registerCleanupFunction(UrlClassifierTestUtils.cleanupTestTrackers);
  await UrlClassifierTestUtils.addTestTrackers();

  // Load the test from a URL that's NOT on the whitelist with tracking protection disabled
  await SpecialPowers.setBoolPref("privacy.trackingprotection.enabled", false);
  await testOnWindow(contentPage2).then(function(aWindow) {
    checkLoads(aWindow, false, false);
    aWindow.close();
  });
  await SpecialPowers.setBoolPref("privacy.trackingprotection.enabled", true);

  // Load the test from a URL that's NOT on the whitelist
  await testOnWindow(contentPage2).then(function(aWindow) {
    checkLoads(aWindow, false, true);
    aWindow.close();
  });

  // Load the test from a URL on the whitelist
  await testOnWindow(contentPage1).then(function(aWindow) {
    checkLoads(aWindow, true, true);
    aWindow.close();
  });

  // Load the test from a URL on the whitelist but without the whitelist
  await SpecialPowers.setCharPref("urlclassifier.trackingWhitelistTable", "");
  await testOnWindow(contentPage1).then(function(aWindow) {
    checkLoads(aWindow, false, true);
    aWindow.close();
  });
  await SpecialPowers.clearUserPref("urlclassifier.trackingWhitelistTable");
  await SpecialPowers.clearUserPref("privacy.trackingprotection.enabled");

  SimpleTest.finish();
}

SimpleTest.waitForExplicitFinish();

</script>

</pre>
<iframe id="testFrame" width="100%" height="100%" onload=""></iframe>
</body>
</html>
